==> code/gwl/02_index.Rmd <==
---
title: "Groundwater level monitoring wells"
output: 
  flexdashboard::flex_dashboard:
    orientation: columns
    vertical_layout: fill
    source_code: "https://github.com/larrywalkerassociates/sites/tree/main/code/gwl"
params: 
  AOI: "000"
---

<style>
  .custom-icon {
    fill: #ADADAD;
  }
  .custom-icon:hover {
    fill: #DA4567;
  }
  div.leaflet-popup-content {width:auto !important;}
</style>

```{r setup, include=FALSE}
library(flexdashboard)
library(tidyverse)
library(sp)
library(sf)
library(leaflet)
library(leafpop)
library(plotly)
library(htmltools)
library(htmlwidgets)
library(leafem)
library(here)

zipfile <- paste0("https://github.com/larrywalkerassociates/sites/raw/main/data/", params$AOI, "/", params$AOI,"_data.zip")
```

Column Column {.tabset}
-----------------------------------------------------------------------

### Map

```{r}
# read_rds(here("out", "leaf.rds"))
l
```

### Table

```{r}
# read_rds(here("out", "dt.rds")) %>% 
ns %>% 
  select(-lab) %>% 
  DT::datatable(rownames = FALSE, 
                options = list(
                  pageLength = 15, 
                  buttons = c('copy', 'csv', 'excel'),
                  escape = FALSE)
                ) 
```

### Data

<h4>**Download data**</h4>

`r paste0('<a href=', '"', zipfile, '"', 'target = "_blank">')`
	<svg height = "50" aria-hidden="true" focusable="false" data-prefix="fas" data-icon="download" class="svg-inline--fa fa-download fa-w-16 custom-icon" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path  d="M216 0h80c13.3 0 24 10.7 24 24v168h87.7c17.8 0 26.7 21.5 14.1 34.1L269.7 378.3c-7.5 7.5-19.8 7.5-27.3 0L90.1 226.1c-12.6-12.6-3.7-34.1 14.1-34.1H192V24c0-13.3 10.7-24 24-24zm296 376v112c0 13.3-10.7 24-24 24H24c-13.3 0-24-10.7-24-24V376c0-13.3 10.7-24 24-24h146.7l49 49c20.1 20.1 52.5 20.1 72.6 0l49-49H488c13.3 0 24 10.7 24 24zm-124 88c0-11-9-20-20-20s-20 9-20 20 9 20 20 20 20-9 20-20zm64 0c0-11-9-20-20-20s-20 9-20 20 9 20 20 20 20-9 20-20z"></path></svg>
</a>

Monitoring data for all wells is given in a zipfile with separate csv files per well named by unique `SITE_CODE` parameter (`SITE_CODE.csv`), and one long-format table with all wells combined (`all_gwl_data.csv`).  

No crosswalk between `SITE_CODE` or `SWN` and `WCRNumber` exists to link these data to OSWCR, so finding well completion reports for monitoring wells identified in this dataset requires reaching out to the primary monitoring agency.  

Data comes from the roughly 1.8 million records in the [Periodic Groundwater Level Measurements database](https://data.ca.gov/dataset/periodic-groundwater-level-measurements), which is regularly maintained and supersedes CASGEM.
